+++ CHANGELOG.md
@@ -6,4 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 
+## [1.0.0] - 2024-05-09
+### Added
+- `Getset` procedural macro.
+- `Getset` procedural macro field attributes:
+  - `get`
+  - `get_mut`
+  - `get_copy`
+  - `get_deref`
+  - `get_deref_mut`
+  - `get_deref_copy`
+  - `get_as_ref`
+  - `get_as_deref`
+  - `get_as_deref_mut`
+  - `set`
+
 <!-- next-url -->
+[1.0.0]: https://github.com/andrewsonin/gset/releases/tag/v1.0.0